% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rtf_files.R
\name{merge_rtf_files}
\alias{merge_rtf_files}
\title{Merge RTF files into one single RTF file}
\usage{
merge_rtf_files(path)
}
\arguments{
\item{path}{a vector of RTF file path}
}
\description{
Merge RTF files into one single RTF file
}
