% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient_rtf.R
\name{orient_rtf}
\alias{orient_rtf}
\title{rtf_text}
\usage{
orient_rtf(rtf_text, landscape = TRUE)
}
\arguments{
\item{rtf_text}{A vector of character lines from an RTF.}
}
\value{
A character vector containing lines from RTF
}
\description{
This orients a RTF table, in the form of a vector
where each element is a line from an RTF table, from
landscape to portrait or the portrait to landscape
}
\examples{
rtflines = readLines("rtf_file.rtf")
turned_rtf = orient_rtf(rtflines, landscape)
writeLines(turned_rtf, "turned_rtf.rtf")

}
